+++ src/app/[variants]/(main)/home/_layout/Body/Project/List/Editing.tsx
@@ -1,7 +1,7 @@
 import { Input, Popover } from '@lobehub/ui';
 import { memo, useCallback, useState } from 'react';
 
-import { useKnowledgeBaseStore } from '@/store/knowledgeBase';
+import { useKnowledgeBaseStore } from '@/store/library';
 
 interface EditingProps {
   id: string;
