+++ src/app/[variants]/(main)/home/_layout/Body/Project/List/index.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from 'react-i18next';
 import { Link, useNavigate } from 'react-router-dom';
 
 import { LIBRARY_URL } from '@/const/url';
-import { useKnowledgeBaseStore } from '@/store/knowledgeBase';
+import { useKnowledgeBaseStore } from '@/store/library';
 
 import EmptyNavItem from '../../../../../../../../features/NavPanel/components/EmptyNavItem';
 import SkeletonList from '../../../../../../../../features/NavPanel/components/SkeletonList';
