+++ src/app/[variants]/(main)/home/_layout/Body/Project/List/useDropdownMenu.tsx
@@ -4,7 +4,7 @@ import { PencilLine, Trash } from 'lucide-react';
 import { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { useKnowledgeBaseStore } from '@/store/knowledgeBase';
+import { useKnowledgeBaseStore } from '@/store/library';
 
 interface ProjectItemDropdownMenuProps {
   id: string;
