+++ src/app/[variants]/(main)/resource/(home)/_layout/Body/LibraryList/List/index.tsx
@@ -3,7 +3,7 @@
 import { Flexbox } from '@lobehub/ui';
 import { memo } from 'react';
 
-import { useKnowledgeBaseStore } from '@/store/knowledgeBase';
+import { useKnowledgeBaseStore } from '@/store/library';
 
 import Item from './Item';
