+++ src/app/[variants]/(main)/resource/(home)/_layout/Body/LibraryList/index.tsx
@@ -7,7 +7,7 @@ import { useNavigate } from 'react-router-dom';
 import { useCreateNewModal } from '@/features/LibraryModal';
 import EmptyNavItem from '@/features/NavPanel/components/EmptyNavItem';
 import SkeletonList from '@/features/NavPanel/components/SkeletonList';
-import { useKnowledgeBaseStore } from '@/store/knowledgeBase';
+import { useKnowledgeBaseStore } from '@/store/library';
 
 import List from './List';
