+++ src/app/[variants]/(main)/resource/features/hooks/useKnowledgeItem.ts
@@ -1,4 +1,4 @@
-import { useKnowledgeBaseStore } from '@/store/knowledgeBase';
+import { useKnowledgeBaseStore } from '@/store/library';
 
 export const useKnowledgeBaseItem = (id: string) => {
   const useFetchKnowledgeBaseItem = useKnowledgeBaseStore((s) => s.useFetchKnowledgeBaseItem);
